requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
openai
chromadb
PyPDF2
python-dotenv
pyttsx3
SpeechRecognition